carla/requirements.txt
absl-py==0.15.0
albumentations==1.2.1
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==3.3.1
cached-property==1.5.2
cachetools==4.2.4
carla-birdeye-view==1.1.1
catkin-pkg==0.5.2
catkin-tools==0.8.5
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
clang==5.0
cycler==0.10.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
docutils==0.19
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.23.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==2.10
imageio==2.15.0
importlib-metadata==6.0.0
ipykernel==5.5.6
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==2.11.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.4
notebook==6.2.0
numpy==1.21.6
oauthlib==3.1.0
open3d==0.10.0.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
osrf-pycommon==2.0.2
packaging==20.9
pandas==0.25.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==2.0.10
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
Pygments==2.11.2
pyparsing==3.0.9
pyrsistent==0.17.3
python-dateutil==2.8.2
python-version==0.0.2
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==20.0.0
qudida==0.0.4
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.7.2
schedule==1.1.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tornado==6.1
traitlets==4.3.3
typing_extensions==4.4.0
urllib3==1.26.9
vcstool==0.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.1
wrapt==1.12.1
zipp==3.12.0
